🛠 :: (#409) 코멘트 반영

+++ jobis-application/src/main/java/team/retum/jobis/domain/recruitment/usecase/StudentQueryRecruitmentsUseCase.java
@@ -24,7 +24,7 @@ public class StudentQueryRecruitmentsUseCase {
 
     public StudentQueryRecruitmentsResponse execute(
             String name,
-            Integer page,
+            Long page,
             Long jobCode,
             List<Long> codeIds
     ) {
